requirements.txt
flask~=3.0.3
flask-login~=0.6.3
flask_sqlalchemy
flask_session
flask_wtf
redis~=5.1.1
azure-core~=1.31.0
azure.identity
azure.keyvault
python-dotenv~=1.0.1
werkzeug~=3.0.4
email_validator